Fixing the warnings (#234)
* conditional KPI

only for VELEC, the KPI parameters are changing when the throttle is released.

* Delete RTX_Config.c.base@5.1.1

* Delete RTX_Config.h.base@5.5.2

* Delete startup_RA.c

* Delete startup_RA.c.base@1.0.0

* Delete system_RA.c

* Delete system_RA.c.base@1.0.0

* Delete startup_RA.c.base@1.0.0

* Delete system_RA.c.base@1.0.0

* comment added

* maximizing torque and power

the max. torque was increased to 1500 and max power was set to 1500W

* Revert "Merge branch 'main' of https://github.com/tristar-multicopters/GNR_firmware"

This reverts commit d75e7e30d844d6e7d84c2a98ba9c06395b26cef0, reversing
changes made to 9d386becc5ad6a9925ccb5b81e6560e443580c2e.

* fixing the warning

all the warning related to the last motor control commit was fixed

* fixing_warnings

---------

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/inc/application_motor/parameters_conversion.h
@@ -73,12 +73,12 @@
 /* The maximum current from Current Sensor with AMPLIFICATION_GAIN that we can measure using ADC with ADC_REFERENCE_VOLTAGE */
 #define current_correction_factor   0.85  /* correction factor for the current sensor range */
 #define MAX_MEASURABLE_CURRENT  (ADC_REFERENCE_VOLTAGE/(2*AMPLIFICATION_GAIN)* current_correction_factor)  
-#define NOMINAL_PEAK_CURRENT    (PEAK_CURRENT_amps * 65535 / (2 * MAX_MEASURABLE_CURRENT))   /* Maximum current amplitude that can be injected per phase in digital Amps */
-#define ID_DEMAG        (ID_DEMAG_amps  * 65535 / (2 * MAX_MEASURABLE_CURRENT))
+#define NOMINAL_PEAK_CURRENT    (uint16_t)(PEAK_CURRENT_amps * 65535 / (2 * MAX_MEASURABLE_CURRENT))   /* Maximum current amplitude that can be injected per phase in digital Amps */
+#define ID_DEMAG        (uint16_t)(ID_DEMAG_amps  * 65535 / (2 * MAX_MEASURABLE_CURRENT))
 
-#define OCSP_SAFETY_MARGIN 	            (OCSP_SAFETY_MARGIN_amps  * 65535 / (2 * MAX_MEASURABLE_CURRENT))	/* Measured current amplitude can be until SOCP_SAFETY_MARGIN higher
+#define OCSP_SAFETY_MARGIN 	            (uint16_t)(OCSP_SAFETY_MARGIN_amps  * 65535 / (2 * MAX_MEASURABLE_CURRENT))	/* Measured current amplitude can be until SOCP_SAFETY_MARGIN higher
                                                 than reference current before overcurrent software protection triggers */
-#define OCSP_MAX_CURRENT                (OCSP_MAX_CURRENT_amps  * 65535 / (2 * MAX_MEASURABLE_CURRENT))   /* Max current that can be reached before triggering software overcurrent */
+#define OCSP_MAX_CURRENT                (uint16_t)(OCSP_MAX_CURRENT_amps  * 65535 / (2 * MAX_MEASURABLE_CURRENT))   /* Max current that can be reached before triggering software overcurrent */
 
 #define OBS_MINIMUM_SPEED_UNIT    (uint16_t) ((OBS_MINIMUM_SPEED_RPM*SPEED_UNIT)/_RPM)
